feat: CE Phase 2 — render router fold + last /api dataId shims
- app/render/routes.py: 5 bare render routes moved verbatim out of main.py
  (980 -> ~150 lines); router mounts non-CE only, so CE deny-by-default is
  purely structural (middleware backstop remains; per-handler guards gone)
- /api/jwstdata/{id}/preview|pixeldata|cubeinfo|histogram shims (camelCase
  query surface, X-Cube-* header forwarding); histogram was MISSING from
  the Phase 1 inventory though ImageViewer fetches it unconditionally —
  review catch, allowlist doc corrected
- /api/analysis/table-info|table-data|spectral-data (dataId resolution)
- shared app/db/access.py resolve_public_path (replaces cross-module
  private import)
- stale MAX_FITS_ARRAY_ELEMENTS docs corrected (100M -> 200M, pre-existing)

Move fidelity verified mechanically vs origin/main. Live smoke: all shims
200 with real data, bare routes 404 in CE. 16 new tests; 1511 passed.
