+++ processing-engine/app/render/__init__.py
@@ -1,0 +1,1 @@
+"""Render routes package (ADR 0001 migration)."""
